docker-files/worker/requirements.txt
async-timeout==4.0.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.2
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
gurobipy==10.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
python-dateutil==2.8.2
pytz==2022.7.1
rectangle-packer==2.0.1
redis==4.4.2
scikit-learn==1.2.1
scikit-learn-extra==0.2.0
scipy==1.10.0
shapely==2.0.1
six==1.16.0
svgwrite==1.4.3
threadpoolctl==3.1.0
typing-extensions==4.6.3
tzdata==2023.3
Werkzeug==2.2.2
WTForms==3.0.1